feat: Add a separate /rtcstats endpoint.
