+++ jvb/src/main/kotlin/org/jitsi/videobridge/DebugStateMode.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright @ 2020 - present 8x8, Inc.
+ * Copyright @ 2025 - present 8x8, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.videobridge.rest
+package org.jitsi.videobridge
 
-enum class RestApis {
-    COLIBRI,
-    DEBUG,
-    DRAIN,
-    HEALTH,
-    PROMETHEUS,
-    RTCSTATS,
-    SHUTDOWN,
-    VERSION
+enum class DebugStateMode {
+    SHORT,
+    FULL,
+    RTCSTATS
 }
