+++ jvb/src/main/resources/reference.conf
@@ -128,20 +128,23 @@ videobridge {
     debug {
       enabled = true
     }
+    drain {
+      enabled = true
+    }
     health {
+      enabled = true
+    }
+    prometheus {
+      enabled = true
+    }
+    rtcstats {
       enabled = true
     }
     shutdown {
       # Note that the shutdown API requires the COLIBRI API to also be enabled.
       enabled = false
     }
-    drain {
-      enabled = true
-    }
     version {
-      enabled = true
-    }
-    prometheus {
       enabled = true
     }
   }
