+++ jvb/src/test/kotlin/org/jitsi/videobridge/VideobridgeTest.kt
@@ -51,7 +51,7 @@ class VideobridgeTest : ShouldSpec() {
 
     init {
         context("Debug state should be JSON") {
-            videobridge.getDebugState(null, null, true).shouldBeValidJson()
+            videobridge.getDebugState(null, null, DebugStateMode.FULL).shouldBeValidJson()
         }
         context("Shutdown") {
             context("when a conference is active") {
